+++ lib/simctl/command.rb
@@ -13,6 +13,7 @@ require 'simctl/command/reset'
 require 'simctl/command/shutdown'
 require 'simctl/command/spawn'
 require 'simctl/command/uninstall'
+require 'simctl/command/warmup'
 require 'simctl/executor'
 require 'shellwords'
 
@@ -35,6 +36,7 @@ module SimCtl
     include SimCtl::Command::Shutdown
     include SimCtl::Command::Spawn
     include SimCtl::Command::Uninstall
+    include SimCtl::Command::Warmup
 
     def device_set_path=(device_set_path)
       @device_set_path = File.expand_path(device_set_path)